:rocket: (NextAuth): Upgraded to NextAuth v5
Using nextauth v5, added a middleware file, necessary nextauth changes and some UI enhancements to tell the user why they can't access the /calendar page
