+++ app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,2 @@
+export { GET, POST } from '@/auth';
+export const runtime = 'edge';
