+++ app/layout.tsx
@@ -2,10 +2,10 @@ import type { Metadata } from 'next';
 import { Inter, Poppins } from 'next/font/google';
 
 import './globals.css';
-import { authOptions } from '@/pages/api/auth/[...nextauth]';
-import { getServerSession } from 'next-auth';
+import { auth } from '@/auth';
 import SessionProvider from './SessionProvider';
 import Navbar from '@/components/Navbar/Navbar';
+import { Toaster } from '@/components/ui/sonner';
 
 const inter = Inter({ subsets: ['latin'] });
 const poppins = Poppins({
@@ -26,13 +26,14 @@ export default async function RootLayout({
     children: React.ReactNode;
 }) {
     console.log('Children:', children); // Logging children variable
-    const session = await getServerSession(authOptions);
+    const session = await auth();
     return (
         <html lang="en">
             <body className={`${poppins.variable} ${inter.className}`}>
                 <SessionProvider session={session}>
                     <Navbar />
                     {children}
+                    <Toaster />
                 </SessionProvider>
             </body>
         </html>
