+++ package.json
@@ -26,7 +26,8 @@
         "framer-motion": "^11.0.8",
         "lucide-react": "^0.325.0",
         "next": "14.1.0",
-        "next-auth": "^4.24.5",
+        "next-auth": "^5.0.0-beta.16",
+        "next-themes": "^0.3.0",
         "react": "^18",
         "react-day-picker": "^8.10.0",
         "react-dnd": "^16.0.1",
@@ -35,6 +36,7 @@
         "react-hook-form": "^7.50.1",
         "react-icons": "^5.0.1",
         "react-resizable-panels": "^2.0.9",
+        "sonner": "^1.4.41",
         "tailwind-merge": "^2.2.1",
         "tailwindcss-animate": "^1.0.7"
     },
