+++ tsconfig.json
@@ -25,8 +25,8 @@
         "next-env.d.ts",
         "**/*.ts",
         "**/*.tsx",
-        ".next/types/**/*.ts", "pages/api/auth/[...nextauth].ts",
-        "pages/api/auth/[...nextauth].js"
+        ".next/types/**/*.ts", "app/api/auth/[...nextauth].ts",
+        "app/api/auth/[...nextauth].js"
     ],
     "exclude": ["node_modules"]
 }
